services/ingestion/requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
pydantic-settings>=2.1.0

# HTTP & Networking
httpx>=0.26.0
python-multipart>=0.0.6
aiofiles>=23.2.1

# Video Processing
opencv-python>=4.9.0.80
pillow>=10.3.0
python-magic-bin>=0.4.14

# Computer Vision Models
torch>=2.1.0
torchvision>=0.16.0
kornia>=0.7.1
scikit-image>=0.22.0
git+https://github.com/facebookresearch/segment-anything.git
ever-beta>=0.2.0
git+https://github.com/Z-Zheng/pytorch-change-models.git

# Storage
boto3>=1.34.34
minio>=7.2.3

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.25
alembic>=1.13.1
asyncpg>=0.29.0

# Redis
redis>=5.0.1

# Utilities
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Monitoring & Logging
prometheus-client>=0.19.0
structlog>=24.1.0

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0
faker>=22.5.0
